+++ cardinal-components-world/src/main/java/org/ladysnake/cca/internal/world/CardinalComponentsWorld.java
@@ -26,14 +26,22 @@ import com.mojang.datafixers.util.Unit;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerWorldEvents;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.network.packet.CustomPayload;
+import net.minecraft.registry.RegistryKey;
+import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
+import org.ladysnake.cca.api.v3.component.ComponentContainer;
 import org.ladysnake.cca.api.v3.component.ComponentKey;
 import org.ladysnake.cca.api.v3.component.ComponentProvider;
 import org.ladysnake.cca.api.v3.world.WorldSyncCallback;
 import org.ladysnake.cca.internal.base.ComponentUpdatePayload;
 import org.ladysnake.cca.internal.base.MorePacketCodecs;
 
+import java.util.IdentityHashMap;
+import java.util.Map;
+
 public final class CardinalComponentsWorld {
     public static final CustomPayload.Id<ComponentUpdatePayload<Unit>> PACKET_ID = ComponentUpdatePayload.id("world_sync");
+    private static final Map<@Nullable RegistryKey<World>, ComponentContainer.Factory<World>> worldContainerFactories = new IdentityHashMap<>();
 
     public static void init() {
         if (FabricLoader.getInstance().isModLoaded("fabric-networking-api-v1")) {
@@ -48,6 +56,26 @@ public final class CardinalComponentsWorld {
         if (FabricLoader.getInstance().isModLoaded("fabric-lifecycle-events-v1")) {
             ServerWorldEvents.LOAD.register((server, world) -> ((ComponentProvider) world).getComponentContainer().onServerLoad());
             ServerWorldEvents.UNLOAD.register((server, world) -> ((ComponentProvider) world).getComponentContainer().onServerUnload());
+        }
+    }
+
+    public static ComponentContainer createComponents(World world) {
+        return worldContainerFactories.computeIfAbsent(
+            world.getRegistryKey(),
+            CardinalComponentsWorld::getWorldComponentFactory
+        ).createContainer(world);
+    }
+
+
+    private static synchronized ComponentContainer.Factory<World> getWorldComponentFactory(RegistryKey<World> dimensionKey) {
+        // need to check again despite synchronization, because the factory may have been generated while waiting from createComponents
+        ComponentContainer.Factory<World> existing = worldContainerFactories.get(dimensionKey);
+        if (existing != null) return existing;
+
+        if (StaticWorldComponentPlugin.INSTANCE.requiresStaticFactory(dimensionKey)) {
+            return StaticWorldComponentPlugin.INSTANCE.buildDedicatedFactory(dimensionKey);
+        } else {
+            return worldContainerFactories.computeIfAbsent(null, n -> StaticWorldComponentPlugin.INSTANCE.buildDedicatedFactory(null));
         }
     }
 }